+++ scripts/measure-navlyn-performance.ps1
@@ -420,7 +420,7 @@ function Invoke-McpToolScenario {
                 capabilities = @{}
                 clientInfo = @{
                     name = 'navlyn-performance'
-                    version = '0.1.0'
+                    version = '0.2.0'
                 }
             }
         }